Refactor: PointItem으로 컴포넌트명 변경
