+++ components/PointItem.tsx
